# api/openapi.yaml
openapi: 3.1.0
info:
  title: go2rtc
  license: { name: MIT,url: https://opensource.org/licenses/MIT }
  version: 1.0.0
  contact: { url: https://github.com/dadav/go2rtc }
  description: |
    Ultimate camera streaming application with support RTSP, RTMP, HTTP-FLV, WebRTC, MSE, HLS, MP4, MJPEG, HomeKit, FFmpeg, etc.
servers:
  - url: http://localhost:1984
components:
  parameters:
    stream_src_path:
      name: src
      in: path
      description: Source stream name
      required: true
      schema: { type: string }
      example: camera1
...
paths:
...
  /api/exit:
    post:
      summary: Close application
      tags: [ Application ]
      parameters:
        - name: code
          in: query
          description: Application exit code
          required: false
          schema: { type: integer }
          example: 100
      responses:
            default:
              description: Default response
...
  /api/config:
    get:
      summary: Get main config file content
      tags: [ Config ]
      responses:
        "200":
          description: ""
          content:
            application/yaml: { example: "streams:..." }
    post:
      summary: Rewrite main config file
      tags: [ Config ]
      requestBody:
        content:
          "*/*": { example: "streams:..." }
      responses:
            default:
              description: Default response
    patch:
      summary: Merge changes to main config file
      tags: [ Config ]
      requestBody:
        content:
          "*/*": { example: "streams:..." }
      responses:
            default:
              description: Default response
  /api/streams:
    get:
      summary: Get all streams info
      tags: [ Streams list ]
      responses:
        "200":
          description: ""
          content:
            application/json: { example: { camera1: { producers: [ ],consumers: [ ] } } }
    put:
      summary: Create new stream
      tags: [ Streams list ]
      parameters:
        - name: src
          in: query
          description: Stream source (URI)
          required: true
          schema: { type: string }
          example: "rtsp://rtsp:12345678@192.168.1.123/av_stream/ch0"
        - name: name
          in: query
          description: Stream name
          required: false
          schema: { type: string }
          example: camera1
      responses:
            default:
              description: Default response
    patch:
      summary: Update stream source
      tags: [ Streams list ]
      parameters:
        - name: src
          in: query
          description: Stream source (URI)
          required: true
          schema: { type: string }
          example: "rtsp://rtsp:12345678@192.168.1.123/av_stream/ch0"
        - name: name
          in: query
          description: Stream name
          required: true
          schema: { type: string }
          example: camera1
      responses:
            default:
              description: Default response
    delete:
      summary: Delete stream
      tags: [ Streams list ]
      parameters:
        - name: src
          in: query
          description: Stream name
          required: true
          schema: { type: string }
          example: camera1
      responses:
            default:
              description: Default response
    post:
      summary: Send stream from source to destination
      description: "[Stream to camera](https://github.com/dadav/go2rtc#stream-to-camera)"
      tags: [ Streams list ]
      parameters:
        - name: src
          in: query
          description: Stream source (URI)
          required: true
          schema: { type: string }
          example: "ffmpeg:http://example.com/song.mp3#audio=pcma#input=file"
        - name: dst
          in: query
          description: Destination stream name
          required: true
          schema: { type: string }
          example: camera1
      responses:
            default:
              description: Default response
  /api/streams?src={src}:
    get:
      summary: Get stream info in JSON format
      tags: [ Consume stream ]
      parameters:
        - $ref: "#/components/parameters/stream_src_path"
      responses:
        "200":
          description: ""
          content:
            application/json:
              example: { producers: [ { url: "rtsp://rtsp:12345678@192.168.1.123/av_stream/ch0" } ], consumers: [ ] }
...